refactor(materials): consolidate neo-Hookean Wa/Wb/Wc formulations into one parametrized material each
Wa/Wb/Wc (and their plastic counterparts) were six near-identical
classes differing only in their energy-density/stress/tangent
functions. Replace them with a single NeoHookeanMaterial and a single
NeoHookeanPlasticMaterial, each taking the formulation as the first
material property (1=Wa, 2=Wb, 3=Wc) and dispatching to a small
module-level pure function per formulation. Adding a fourth potential
now means adding one function + one dict entry in one file, instead of
a whole new class/file pair.

- edelweissfe/materials/neohooke/neohooke.py replaces
  neohookepencegouformulationa/b/c.py + pencegoubase.py.
- edelweissfe/materials/neohookeplastic/neohookeplastic.py replaces
  neohookepencegouformulationaplastic/bplastic/cplastic.py +
  pencegouplasticbase.py.
- materiallibrary.py: "neohookewa"/"neohookewb"/"neohookewc" collapse
  to "neohooke"; "neohookewaplastic"/"wbplastic"/"wcplastic" collapse
  to "neohookeplastic". Existing *.inp files updated to the new
  material name and formulation-selector property.
- BaseHyperElasticMaterial.computePlaneKirchhoff is no longer
  @abstractmethod (mirrors the computePlaneStress fix from the
  previous commit) — the new consolidated classes have no reason to
  carry a no-op super()-delegating override just to satisfy ABC
  instantiation.
- doc/source/documentation/materials.rst updated to describe the
  unified materials and the formulation selector.

Verified by direct numerical comparison against the old (pre-refactor)
classes recovered from git history: 300 random deformation gradients
per elastic formulation (stress/tangent/energy bit-for-bit identical),
and a 40-step incremental loading path per plastic formulation that
exercises actual plastic yielding every step (stress/tangent/
energy/kappa bit-for-bit identical after fixing one transcription slip
caught by this comparison: neohooke.py's Wa tangent originally used
muBar=(K/2-mu/3)*(2J^2-J), copied from the plastic Wa's elastic-tangent
formula, instead of the standalone elastic Wa's own
muBar=(K-2/3*mu)*(2J^2-J) — a pre-existing, intentionally-preserved
difference between the two contexts, not something to unify).

Also ran the full testfiles/edelweiss-only/ suite (0 failures),
including all six Wa/Wb/Wc elastic/plastic regression tests now
exercised through the single consolidated classes, plus
CuboidLatticeGenerator (a 3D solid test using formulation Wb).
