--- edelweissfe/materials/neohooke/neohookepencegouformulationb.py
@@ -1,87 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-#  ---------------------------------------------------------------------
-#
-#  _____    _      _              _         _____ _____
-# | ____|__| | ___| |_      _____(_)___ ___|  ___| ____|
-# |  _| / _` |/ _ \ \ \ /\ / / _ \ / __/ __| |_  |  _|
-# | |__| (_| |  __/ |\ V  V /  __/ \__ \__ \  _| | |___
-# |_____\__,_|\___|_| \_/\_/ \___|_|___/___/_|   |_____|
-#
-#
-#  Unit of Strength of Materials and Structural Analysis
-#  University of Innsbruck,
-#  2017 - today
-#
-#  Daniel Reitmair daniel.reitmair@uibk.ac.at
-#
-#  This file is part of EdelweissFE.
-#
-#  This library is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 2.1 of the License, or (at your option) any later version.
-#
-#  The full text of the license can be found in the file LICENSE.md at
-#  the top level directory of EdelweissFE.
-#  ---------------------------------------------------------------------
-
-import numpy as np
-import numpy.linalg as lin
-
-from edelweissfe.materials.neohooke.pencegoubase import NeoHookeanPenceGouBaseMaterial
-from edelweissfe.utils.voigtnotation import doVoigtStress
-
-
-class NeoHookeanWbMaterial(NeoHookeanPenceGouBaseMaterial):
-    """Neo-Hookean Wb material according to [1] with the following energy density function.
-
-    W_b (I1, J) = mu/2 * (I1/J^(2/3) - 3) + K/8 * (J² + 1/J² - 2).
-
-    [1] Pence, T. J., & Gou, K. (2015). On compressible versions of the incompressible neo-Hookean material.
-        Mathematics and Mechanics of Solids, 20(2), 157–182. https://doi.org/10.1177/1081286514544258
-
-    Parameters
-    ----------
-    materialProperties
-        The numpy array containing the material properties for the requested material."""
-
-    def computeKirchhoff(
-        self,
-        stress: np.ndarray,
-        dStress_dDeformationGradient: np.ndarray,
-        deformationGradient: np.ndarray,
-        time: float,
-        dTime: float,
-    ):
-        """Computes the stresses for a 3D material/2D material with plane strain.
-
-        Parameters
-        ----------
-        stress
-            Vector containing the stresses.
-        dStress_dDeformationGradient
-            Matrix containing dStress/dStrain.
-        deformationGradient
-            The deformation gradient at time step t.
-        time
-            Array of step time and total time.
-        dTime
-            Current time step size."""
-
-        F = deformationGradient
-        B = F @ F.T  # left Cauchy-Green tensor
-        invF = lin.inv(F)  # inverse
-        J = lin.det(F)
-        I1 = np.trace(B)
-        lambdaHat = self._K / 2 * (J**2 + 1 / J**2)
-        muBar = self._mu / (3 * J ** (2 / 3))
-        lambdaBar = self._K / 4 * (J**2 - 1 / J**2) - muBar * I1
-        stress[:] = doVoigtStress(3, self._mu / J ** (2 / 3) * B + lambdaBar * np.eye(3))
-        dStress_dDeformationGradient[:] = (
-            3 * muBar * (np.einsum("ik,jl->ijkl", np.eye(3), F) + np.einsum("il,jk->ijkl", F, np.eye(3)))
-            - 2 * muBar * np.einsum("ij,lk->ijkl", B, invF)
-            + (lambdaHat + 2 / 3 * I1 * muBar) * np.einsum("ij,lk->ijkl", np.eye(3), invF)
-            - 2 * muBar * np.einsum("ij,kl->ijkl", np.eye(3), F)
-        )
-        self._energy[0] = self._mu / 2 * (I1 / J ** (2 / 3) - 3) + self._K / 8 * (J**2 + 1 / J**2 - 2)
