+++ testfiles/edelweiss-only/CantileverBeamQuad8NeoHookeWc/test.inp
@@ -1,5 +1,5 @@
-*material, name=neohookewc, id=neohookewc, provider=edelweiss
-91304.34783, 100000.
+*material, name=neohooke, id=neohookewc, provider=edelweiss
+3, 91304.34783, 100000.
 
 *section, name=section1, thickness=40.0, material=neohookewc, type=plane
 all
